refactor: utils 폴더 생성 및 이동
utilModule은 만들지 않았습니다.

- #199
