+++ src/auth/auth.module.ts
@@ -4,7 +4,7 @@ import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserInfo } from 'src/entities/user-info.entity';
-import { GoogleApiModule } from 'src/google-api/google-api.module';
+import { GoogleApiModule } from 'src/utils/google-api/google-api.module';
 import { Auth42Controller } from './auth.controller';
 import { AuthService } from './auth.service';
 import { AuthRepository } from './repository/auth.repository';
