+++ src/tag-log/tag-log.module.ts
@@ -1,18 +1,18 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { DateCalculatorModule } from 'src/data-calculator/data-calculator.module';
 import { DeviceInfo } from 'src/entities/device-info.entity';
 import { PairInfo } from 'src/entities/pair-info.entity';
 import { TagLog } from 'src/entities/tag-log.entity';
 import { StatisticsModule } from 'src/statistics/statictics.module';
 import { UserModule } from 'src/user/user.module';
+import { DateCalculatorModule } from 'src/utils/data-calculator/data-calculator.module';
 import { DeviceInfoRepository } from './repository/mysql/device-info.repository';
 import { PairInfoRepository } from './repository/mysql/pair-info.repository';
 import { TagLogRepository } from './repository/mysql/tag-log.repository';
 import { TagLogAdminController } from './tag-log-admin.controller';
 import { TagLogAdminService } from './tag-log-admin.service';
-import { TagLogService } from './tag-log.service';
 import { TagLogController } from './tag-log.controller';
+import { TagLogService } from './tag-log.service';
 
 const tagLogRepo = {
   provide: 'ITagLogRepository',
