+++ src/utils/data-calculator/common.constants.ts
